Support grouping embeds on a single line
This is WIP implementation of
https://github.com/fleather-editor/fleather/discussions/62 .

Atm it works without introducing a new embed type. I like the simplicity
of it but it might not be the case in real life scenarios (e.g
horizontal rule must be the only child on a line). So maybe we can add a
constraint on horizontal rule that it must be lone child.

How does it work?

I have implemented `GroupEmbedProxy` (I don't like the name :/). This
new proxy class simply delegates laying out the children to Flutter's
`Flex` widget. I had to create new  `RenderGroupEmbedProxy` class so
that it implements caret methods.

What I need help with?

* Feedback
* Testing
* Pointers where and how to write tests (I am new to Flutter)

What I am planning to work on more?

* Fix: Inserting text between the embeds only splits new line and
doesn't insert text. Text later reappears after inserting new line.
* Fix: It is not possible to remove new line between two embeds.